requirements.txt
annotated-types==0.7.0
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
executing==2.0.0
filelock==3.13.1
fonttools==4.43.1
install==1.3.5
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.1
opencv-python==4.10.0.84
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==25.1.1
scikit-learn==1.5.2
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.12.2
tzdata==2023.3
virtualenv==20.25.0
wcwidth==0.2.8
